% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_annotations.R
\name{align_annotations}
\alias{align_annotations}
\title{Vertical alignment of package annotations}
\usage{
align_annotations(
  annot_string,
 
    regex_ai = "(?!\\\\))(?!\\\\s)(?=\\\\#\\\\sCR|\\\\#\\\\sBio|\\\\#\\\\snot i|\\\\#\\\\s\\\\[)",
  sep_str = ""
)
}
\arguments{
\item{annot_string}{Character vector with annotated package calls.}

\item{regex_ai}{Custom regular expression to match lines with package annotations.}

\item{sep_str}{Whitespace separator.}
}
\value{
A character vector with vertically aligned package calls.
}
\description{
Internal helper function, based on `unheadr::regex_valign``
}
