% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_active_file.R
\name{annotate_active_file}
\alias{annotate_active_file}
\title{Annotate Active File}
\usage{
annotate_active_file()
}
\value{
No return value, wraps [annotate_pkg_calls()] for access via Addin
}
\description{
Annotates package load calls in the active .R file
}
