% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pkg_datasets_active.R
\name{annotate_data_active_file}
\alias{annotate_data_active_file}
\title{Annotate loaded data in active file}
\usage{
annotate_data_active_file()
}
\value{
No return value, wraps [annotate_pkg_datasets()] for access via Addin
}
\description{
Annotates package load calls with loaded datasets in the active .R file
}
