% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_fun_calls.R
\name{annotate_fun_calls}
\alias{annotate_fun_calls}
\title{Annotate function calls}
\usage{
annotate_fun_calls(string_og)
}
\arguments{
\item{string_og}{text string (script) with package load calls}
}
\value{
text string with function call annotations. Will make note of
  packages not currently installed. Lines with existing comments or
  annotations are ignored by the regular expression that matches package
  names. Be aware that package-specific methods (e.g., print or summary
  methods) will be annotated too.
}
\description{
Annotate function calls
}
\examples{
test_string <- c("library(boot)\nrequire(lattice)\ncanonical.theme()")
cat(annotate_fun_calls(test_string))

}
