% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pckg_calls.R
\name{annotate_pckg_calls}
\alias{annotate_pckg_calls}
\title{Annotate package calls}
\usage{
annotate_pckg_calls(string_og, pckg_field = "Title")
}
\arguments{
\item{string_og}{text string (script) with package load calls}

\item{pckg_field}{field from package description to retrieve, defaults to
"Title"}
}
\value{
text string with package Title annotations
#'
}
\description{
Annotate package calls
}
\examples{
test_string <- c("library(boot)\\nrequire(Matrix)")
annotate_pckg_calls(test_string)

}
