% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pkg_calls.R
\name{annotate_pkg_calls}
\alias{annotate_pkg_calls}
\title{Annotate package calls}
\usage{
annotate_pkg_calls(string_og, pkg_field = "Title")
}
\arguments{
\item{string_og}{Text string (script) with package load calls.}

\item{pkg_field}{Field from package description to retrieve, defaults to
"Title"}
}
\value{
Text string with package Title annotations. Will make note of
  packages not currently installed.
#'
}
\description{
Annotate package calls
}
\examples{
test_string <- c("library(boot)\nrequire(tools)")
annotate_pkg_calls(test_string)
}
