% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pkg_datasets.R
\name{annotate_pkg_datasets}
\alias{annotate_pkg_datasets}
\title{Annotate package datasets}
\usage{
annotate_pkg_datasets(string_og)
}
\arguments{
\item{string_og}{text string (script) with package load calls}
}
\value{
text string with annotations for datasets loaded from packages
  explicitly, lazily, or through name-spacing. Will make note of packages not
  currently installed. Lines with existing comments or annotations are
  ignored.
}
\description{
Annotate package datasets
}
\examples{
test_string <- c("library(tidyr)\nlibrary(purrr)\ndata(construction)\nsummary(fish_encounters)")
annotate_pkg_datasets(test_string)

}
