% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_repo_source.R
\name{annotate_repo_source}
\alias{annotate_repo_source}
\title{Annotate repository sources}
\usage{
annotate_repo_source(string_og)
}
\arguments{
\item{string_og}{text string (script) with package load calls}
}
\value{
text string with package repository source annotations. Will make
  note of packages not currently installed. Lines with existing comments or
  annotations are ignored by the regular expression that matches package
  names. Also ignores base packages. Local installs now annotated as such.
}
\description{
Annotate repository sources
}
\examples{
test_string <- c("library(knitr)\nrequire(datasets)")
annotate_repo_source(test_string)
}
