% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_repos_active_file.R
\name{annotate_repos_active_file}
\alias{annotate_repos_active_file}
\title{Annotate Repositories in Active File}
\usage{
annotate_repos_active_file()
}
\description{
Annotates package load calls with repository details in the active .R file
}
