% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_repostitle.R
\name{annotate_repostitle}
\alias{annotate_repostitle}
\title{Annotate package titles and repository sources}
\usage{
annotate_repostitle(string_og)
}
\arguments{
\item{string_og}{Text string (script) with package load calls.}
}
\value{
Text string with package titles and package repository source
  annotations. Will make note of packages not currently installed. Lines with
  existing comments or annotations are ignored by the regular expression that
  matches package names. Also ignores base packages.
}
\description{
Annotate package titles and repository sources
}
\details{
Some annotations may be long, check for possible line breaks
  introduced into your script.
}
\examples{
test_string <- c("library(knitr)\nrequire(datasets)")
annotate_repostitle(test_string)
}
