% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_repostitle_active.R
\name{annotate_repostitle_active}
\alias{annotate_repostitle_active}
\title{Annotate Titles and Repositories in Active File}
\usage{
annotate_repostitle_active()
}
\value{
No return value, wraps [annotate_repostitle()] for access via Addin
}
\description{
Annotates package load calls with package titles and repository details in
the active .R file
}
