% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_rsids.R
\name{annotate_rsids}
\alias{annotate_rsids}
\title{annotate_rsids}
\usage{
annotate_rsids(
  df,
  chr_col = chr,
  pos_col = pos,
  effect_allele_col = EFFECT_ALLELE,
  other_allele_col = OTHER_ALLELE,
  multi_thread = TRUE,
  progress_bar = TRUE,
  ref_dataset = NA
)
}
\arguments{
\item{df}{data frame or tibble containing data for annotation}

\item{chr_col}{column containing chromosome data}

\item{pos_col}{column containing position data}

\item{effect_allele_col}{column containing effect alleles (used only for rsid matching, no allele harmonization is performed)}

\item{other_allele_col}{column containing effect alleles (used only for rsid matching, no allele harmonization is performed)}

\item{multi_thread}{enable multi-threaded processing. Use with caution if resource-limited (default = TRUE)}

\item{progress_bar}{enable progress bar (default = TRUE)}

\item{ref_dataset}{path to folder containing reference dataset in .parquet format, which will be rapidly loaded using the arrow::open_dataset() function}
}
\value{
Tibble containing original dataset + additional column including rsids
}
\description{
Annotates genome wide association study (GWAS) results (or other tabular data including chromosome, position, and allele data) with rsids from a reference dataset.
}
