% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_script.R
\name{annotate_script}
\alias{annotate_script}
\title{Annotate Script}
\usage{
annotate_script(script_file, pckg_field)
}
\arguments{
\item{script_file}{path to an R script}

\item{pckg_field}{field from package description to retrieve, defaults to
"Title"}
}
\value{
currently prints the annotated script to the console
}
\description{
Annotate Script
}
