% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateR-package.R
\docType{package}
\name{annotateR-package}
\alias{annotateR}
\alias{annotateR-package}
\title{annotateR: Annotates GWAS Data from Reference Dataset}
\description{
The goal of annotateR is to enable annotation of genome wide association
    study (GWAS) results (or other tabular data including chromosome,
    position, and allele data) with RSIDs from a reference dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mglev1n/annotateR}
  \item Report bugs at \url{https://github.com/mglev1n/annotateR/issues}
}

}
\author{
\strong{Maintainer}: Michael Levin \email{michael.levin@pennmedicine.upenn.edu} (\href{https://orcid.org/0000-0002-9937-9932}{ORCID})

}
\keyword{internal}
