% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pkg_datasets.R
\name{check_pkgs}
\alias{check_pkgs}
\title{Check if packages are installed}
\usage{
check_pkgs(pkgvec)
}
\arguments{
\item{pkgvec}{Vector of package names}
}
\value{
A data frame with installation status for packages in the input text.
}
\description{
Check if packages are installed
}
