% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_clip_annpckgs.R
\name{eval_clip_annpckgs}
\alias{eval_clip_annpckgs}
\title{Annonate clipboard content}
\usage{
eval_clip_annpckgs()
}
\value{
prints annotated clipboard content to the console
}
\description{
Evaluates and annotates clipboard content.
}
