% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_metapackages.R
\name{expand_metapackages}
\alias{expand_metapackages}
\title{Expand metapackages}
\usage{
expand_metapackages(string_og)
}
\arguments{
\item{string_og}{text string (script) with package load calls}
}
\value{
Text string with metapackage load calls replaced by multiple separate
  calls to the core packages that make up the metapackage. Core packages will
  be fenced in four commenting symbols and the order follows the attachment
  order from each metapackage. Will make note of metapackages not currently
  installed.
}
\description{
Expand metapackages
}
\examples{
test_string <- c("library(tidyverse)")
expand_metapackages(test_string)

}
