% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_metapackages_active.R
\name{expand_metapackages_active_file}
\alias{expand_metapackages_active_file}
\title{Expand metapackages in Active File}
\usage{
expand_metapackages_active_file()
}
\value{
No return value, wraps [expand_metapackages()] for access via Addin
}
\description{
Replaces metapackage load calls in the active .R file with multiple separate
calls to its core packages.
}
