% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pkg_datasets.R
\name{get_pkg_datasets}
\alias{get_pkg_datasets}
\title{Query data from packages}
\usage{
get_pkg_datasets(pkgvec)
}
\arguments{
\item{pkgvec}{Vector of package names}
}
\value{
A data frame with all the bundled data from the specified packages.
}
\description{
Query data from packages
}
\examples{
get_pkg_datasets(c("dplyr","stringr"))

}
