% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pckg_names.R
\name{match_pckg_names}
\alias{match_pckg_names}
\title{Match package names}
\usage{
match_pckg_names(string_og)
}
\arguments{
\item{string_og}{text string (script) with package load calls}
}
\value{
a tibble with the package load calls and package names
#'
}
\description{
Match package names
}
\examples{
test_string <- c("library(boot)\\nrequire(Matrix)")
match_pckg_names(test_string)

}
