% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pkg_names.R
\name{match_pkg_names}
\alias{match_pkg_names}
\title{Match package names}
\usage{
match_pkg_names(string_og)
}
\arguments{
\item{string_og}{Text string (script) with package load calls.}
}
\value{
A tibble with the package load calls and package names.
}
\description{
Match package names
}
\examples{
test_string <- c("library(boot)\nrequire(Matrix)")
match_pkg_names(test_string)
}
