% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_version.R
\name{pkg_version}
\alias{pkg_version}
\title{Retrieve Package version}
\usage{
pkg_version(pkgs_col)
}
\arguments{
\item{pkgs_col}{Package name.}
}
\value{
A character vector with the package version.
}
\description{
Internal helper function.
}
