% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_details.R
\name{repo_details}
\alias{repo_details}
\title{Retrieve repo details}
\usage{
repo_details(pkgs_col)
}
\arguments{
\item{pkgs_col}{Name of variable with the non-CRAN repos.}
}
\value{
A character vector of user names and repository names.
}
\description{
Internal helper function.
}
