% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_fun_calls_active_file.R
\name{annotate_fun_calls_active_file}
\alias{annotate_fun_calls_active_file}
\title{Annotate Function Calls in Active File}
\usage{
annotate_fun_calls_active_file()
}
\value{
No return value, wraps [annotate_fun_calls()] for access via Addin
}
\description{
Annotates package load calls with the used functions of each package in the
active .R file
}
