% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_repo_source.R
\name{annotate_repo_source}
\alias{annotate_repo_source}
\title{Annotate Repo Source}
\usage{
annotate_repo_source(string_og)
}
\arguments{
\item{string_og}{text string (script) with package load calls}
}
\value{
text string with package repository source annotations. Will make note of
  packages not currently installed.
}
\description{
Annotate Repo Source
}
\examples{
#' test_string <- c("library(annotater)\\nrequire(Matrix)")
annotate_repo_source(test_string)
}
