% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_script.R
\name{annotate_script}
\alias{annotate_script}
\title{Annotate Script}
\usage{
annotate_script(script_file, pkg_field = "Title")
}
\arguments{
\item{script_file}{Path to an R script.}

\item{pkg_field}{Field from package description to retrieve, defaults to
"Title".}
}
\value{
Prints the annotated script to the console.
}
\description{
Annotate Script
}
