% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotater-package.R
\docType{package}
\name{annotater-package}
\alias{annotater}
\alias{annotater-package}
\title{annotater: Annotate Package Load Calls}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides non-invasive annotation of package load calls such as \code{library}, \code{p_load}, and \code{require} so that we can have an idea of what the packages we are loading are meant for.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luisDVA/annotater}
  \item \url{https://annotater.liomys.mx}
  \item Report bugs at \url{https://github.com/luisDVA/annotater/issues}
}

}
\author{
\strong{Maintainer}: Luis D. Verde Arregoitia \email{luis@liomys.mx} (\href{https://orcid.org/0000-0001-9520-6543}{ORCID})

}
\keyword{internal}
